# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the HLL App API"
  version: "2.0.0"
  title: "Hyreslandslaget - HLL API"
host: "api.hyreslandslaget.se"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /depots:
    get:
      tags:
      - "Administration"
      summary: "List all available depots"
      description: ""
      operationId: "depotsGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          examples:
            application/json:
            - code: "SE0001"
              name: "Akalla"
              description: "I de gamla tryckertilokalerna på Vandagatan hittar ni\
                \ oss på Depå Akalla."
              phones:
              - number: "08-477 11 50"
                type: "fixed"
              email: "akalla@hllab.se"
              openingHours:
              - name: "Regular"
                openingHours: "09:00-18:00"
              address:
                streetAddress: "Vandagatan 3"
                postalCode: "164 74"
                city: "Akalla"
                country: "SE"
                region: "Stockholm"
                location:
                  latitude: "57.56236"
                  longitude: "18.34526"
              departments:
              - name: "UTHYRNING"
                employees:
                - location: "Manager"
                  firstName: "Niklas"
                  lastName: "Ekwal"
                  phones:
                  - number: "08-477 11 50"
                    type: "fixed"
                  email: "niklas.ekwall@hllab.se"
                  photo: "http://url./niklas.ekwall.png"
          headers: {}
          schema:
            type: "array"
            items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Could not find records."
              code: "404"
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "string"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        503:
          description: "Service Unavailable"
          examples:
            application/json:
              message: "There are no data in the service layer."
              code: "503"
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "string"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Administration"
  /depots/{code}:
    get:
      tags:
      - "Administration"
      summary: "Request single Depo by \"code\" identifier."
      description: ""
      operationId: "depotsCodeGET"
      produces:
      - "application/json"
      parameters:
      - name: "code"
        in: "path"
        description: "The code of the desired depot."
        required: true
        type: "string"
        x-example: "2583743adsf0"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              code: "SE0001"
              name: "Akalla"
              description: "I de gamla tryckertilokalerna på Vandagatan hittar ni\
                \ oss på Depå Akalla."
              phones:
              - number: "08-477 11 50"
                type: "fixed"
              email: "akalla@hllab.se"
              openingHours:
              - name: "Regular"
                openingHours: "09:00-18:00"
              address:
                streetAddress: "Vandagatan 3"
                postalCode: "164 74"
                city: "Akalla"
                country: "SE"
                region: "Stockholm"
                location:
                  latitude: "57.56236"
                  longitude: "18.34526"
              departments:
              - name: "UTHYRNING"
                employees:
                - location: "Manager"
                  firstName: "Niklas"
                  lastName: "Ekwal"
                  phones:
                  - number: "08-477 11 50"
                    type: "fixed"
                  email: "niklas.ekwall@hllab.se"
                  photo: "http://url./niklas.ekwall.png"
          headers: {}
          schema:
            type: "object"
            required:
            - "address"
            - "code"
            - "description"
            - "email"
            - "name"
            - "phones"
            properties:
              code:
                type: "string"
                description: "The Depot identifier"
              name:
                type: "string"
                description: "The name of Depot"
              description:
                type: "string"
                description: "The description of Depot."
              phones:
                type: "array"
                items: {}
              email:
                type: "string"
                description: "The contact mail of Depot."
              openingHours:
                type: "array"
                items: {}
              address:
                $ref: "#/definitions/inline_response_200_1_address"
              departments:
                type: "array"
                description: "The departments in Depot."
                items: {}
        404:
          description: "Not Found"
          examples:
            application/json:
              message: "Could not find record."
              code: "404"
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "string"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
        503:
          description: "Service Unavailable"
          examples:
            application/json:
              message: "There are no data in the service layer."
              code: "503"
              errors:
              - domain: "global"
                reason: "backendError"
                message: "Error message"
          headers: {}
          schema:
            type: "object"
            required:
            - "code"
            - "message"
            properties:
              message:
                type: "string"
                description: "The meessage of Error."
              code:
                type: "string"
                description: "The error code."
              errors:
                type: "array"
                description: "Array of errors."
                items: {}
      x-swagger-router-controller: "Administration"
  /catalogue/groups/{code}:
    get:
      tags:
      - "Administration"
      summary: "Load Group"
      description: ""
      operationId: "catalogueGroupsCodeGET"
      produces:
      - "application/json"
      parameters:
      - name: "code"
        in: "path"
        description: "The group code"
        required: true
        type: "string"
        x-example: "20"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              items:
              - code: "1"
                name: "Heavy machinery"
          headers: {}
          schema:
            type: "object"
            required:
            - "items"
            properties:
              items:
                type: "array"
                description: "List of items"
                items: {}
      x-swagger-router-controller: "Administration"
  /catalogue/products:
    get:
      tags:
      - "Administration"
      summary: "List catalogue assemblies for category"
      description: ""
      operationId: "catalogueProductsGET"
      produces:
      - "application/json"
      parameters:
      - name: "category"
        in: "query"
        description: "The category code"
        required: true
        type: "string"
        x-example: "20"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              items:
              - equipment:
                  article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
                  services:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
                accessories:
                - article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
                    preSelected: true
                  services:
                    article:
                      number: "1234567890"
                      name: "The Dumper"
                      description: "Lorem imposum"
                      sraCode: "211313"
                    offer:
                      name: "\"Hyra\""
                      amount: "123.45"
                      price: "12345 kr / alldager"
                      unit: "st"
          headers: {}
          schema:
            type: "object"
            required:
            - "items"
            properties:
              items:
                type: "array"
                description: "List of items"
                items: {}
      x-swagger-router-controller: "Administration"
  /catalogue/products/{number}:
    get:
      tags:
      - "Administration"
      summary: "Get catalogue assembly"
      description: "Allow user to browse products data"
      operationId: "catalogueProductsNumberGET"
      produces:
      - "application/json"
      parameters:
      - name: "number"
        in: "path"
        required: true
        type: "integer"
      responses:
        200:
          description: "OK"
          examples:
            application/json:
              equipment:
                article:
                  number: "1234567890"
                  name: "The Dumper"
                  description: "Lorem imposum"
                  sraCode: "211313"
                offer:
                  name: "\"Hyra\""
                  amount: "123.45"
                  price: "12345 kr / alldager"
                  unit: "st"
                services:
                  article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
              accessories:
              - article:
                  number: "1234567890"
                  name: "The Dumper"
                  description: "Lorem imposum"
                  sraCode: "211313"
                offer:
                  name: "\"Hyra\""
                  amount: "123.45"
                  price: "12345 kr / alldager"
                  unit: "st"
                  preSelected: true
                services:
                  article:
                    number: "1234567890"
                    name: "The Dumper"
                    description: "Lorem imposum"
                    sraCode: "211313"
                  offer:
                    name: "\"Hyra\""
                    amount: "123.45"
                    price: "12345 kr / alldager"
                    unit: "st"
          headers: {}
          schema:
            type: "object"
            required:
            - "accessories"
            - "equipment"
            properties:
              equipment:
                $ref: "#/definitions/inline_response_200_3_equipment"
              accessories:
                type: "array"
                description: "Accessories related to equipment."
                items: {}
      x-swagger-router-controller: "Administration"
...
definitions:
...
  inline_response_200_1_address_location:
    required:
    - "latitude"
    - "longitude"
    properties:
      latitude:
        type: "string"
        description: "The latitude of the Depot"
      longitude:
        type: "string"
        description: "The longitude of the Depot"
    description: "The lat/long postition of the Depot"
  inline_response_200_1_address:
    required:
    - "city"
    - "country"
    - "location"
    - "postalCode"
    - "region"
    - "streetAddress"
    properties:
      streetAddress:
        type: "string"
        description: "The street name of location."
      postalCode:
        type: "string"
        description: "The postal Code of Location."
      city:
        type: "string"
        description: "The city name of location."
      country:
        type: "string"
        description: "The country name of location."
      region:
        type: "string"
        description: "The region name of location."
      location:
        $ref: "#/definitions/inline_response_200_1_address_location"
    description: "The address of The Depot."
...
  inline_response_200_3_equipment_article:
    required:
    - "number"
    properties:
      number:
        type: "string"
        description: "Article unique number."
      name:
        type: "string"
        description: "Article name long for 30 char."
      description:
        type: "string"
        description: "Article description long for 255 char."
      sraCode:
        type: "string"
        description: "Equipment identifier"
    description: "The Article data"
  inline_response_200_3_equipment_offer:
    required:
    - "amount"
    - "name"
    properties:
      name:
        type: "string"
        description: "Define type on of (Rental,Usage,Purchase)"
      amount:
        type: "string"
        description: "Proposed amount"
      price:
        type: "string"
        description: "Unit price with cost type"
      unit:
        type: "string"
        description: "Unit value"
    description: "The Offer data"
  inline_response_200_3_equipment_services:
    required:
    - "article"
    - "offer"
    properties:
      article:
        $ref: "#/definitions/inline_response_200_3_equipment_article"
      offer:
        $ref: "#/definitions/inline_response_200_3_equipment_offer"
    description: "The services related to product"
  inline_response_200_3_equipment:
    required:
    - "article"
    - "offer"
    properties:
      article:
        $ref: "#/definitions/inline_response_200_3_equipment_article"
      offer:
        $ref: "#/definitions/inline_response_200_3_equipment_offer"
      services:
        $ref: "#/definitions/inline_response_200_3_equipment_services"
    description: "Product for rent with services"
...